Improved Motion Control of Shells
Since the PID controller wasn't working properly (which is weird) I
implemented a simulator that simulates a few paths and chooses the best
one. Not the smoothest of solutions, but it works. It's rather slow (it
works for one shell, but probably won't scale well) but can be improved
using a gradient free optimizer that uses an approximate solution of
your motion equation (which is actually unsolvable) as a starting guess.
However that's a lot of work, so I'm putting it on my to-do list for
now.